Fixed racing game direction control
Due to not bit shifting the car position to the correct direction in the
racing game, it was moving in the oposite direction to what was expected
by the potentiometer control.

Also add some (commented out for now) serial debug lines in various
places, update code formatting, removed some older commented code and
fixed the initial positioning of the paddle X position in Pong.