Case updates and knob for potentionmeter
Add holes in the case to clearly see the charge lights for the LiPo
charger board.

Cut the switch hole up a bit to allow the switch to be installed easier.
The new cutout has a bridge to the other side to allow the upper bit to
be added. This should be removed after printing using a knife.