Change menu animation from scrolling to clicking
The scrolling game selection animation made it difficult to know where
in the menu you are due to the menu screens not having well defined
borders.
This scrolling effect has been replaced with destinct sprites that are
show with every click of the encoder wheel.

This has not been tested with the potentionmeter controller yet!